Migrated most parsing methods to new Parser class
