+++ Basic.Azure.Storage/Communications/BlobService/BlobOperations/GetBlobPropertiesResponse.cs
@@ -42,14 +42,14 @@ namespace Basic.Azure.Storage.Communications.BlobService.BlobOperations
         {
             //TODO: determine what we want to do about potential missing headers and date parsing errors
 
-            LastModified = DateParse.ParseHeader(response.Headers[ProtocolConstants.Headers.LastModified]);
+            LastModified = Parsers.ParseDateHeader(response.Headers[ProtocolConstants.Headers.LastModified]);
             ContentLength = int.Parse(response.Headers[ProtocolConstants.Headers.ContentLength]);
             ContentType = response.Headers[ProtocolConstants.Headers.ContentType];
             ContentMD5 = response.Headers[ProtocolConstants.Headers.ContentMD5];
             ContentEncoding = response.Headers[ProtocolConstants.Headers.ContentEncoding];
             ContentLanguage = response.Headers[ProtocolConstants.Headers.ContentLanguage];
             ETag = response.Headers[ProtocolConstants.Headers.ETag].Trim('"');
-            Date = DateParse.ParseHeader(response.Headers[ProtocolConstants.Headers.OperationDate]);
+            Date = Parsers.ParseDateHeader(response.Headers[ProtocolConstants.Headers.OperationDate]);
 
             ParseBlobType(response);
 
@@ -59,7 +59,7 @@ namespace Basic.Azure.Storage.Communications.BlobService.BlobOperations
 
             ParseLeaseDuration(response);
 
-            Metadata = MetadataParse.ParseMetadata(response);
+            Metadata = Parsers.ParseMetadata(response);
         }
 
         private void ParseBlobType(WebResponse response)
