+++ Basic.Azure.Storage/Communications/BlobService/BlobOperations/PutBlockResponse.cs
@@ -15,7 +15,7 @@ namespace Basic.Azure.Storage.Communications.BlobService.BlobOperations
         {
             //TODO: determine what we want to do about potential missing headers and date parsing errors
 
-            Date = DateParse.ParseHeader(response.Headers[ProtocolConstants.Headers.OperationDate]);
+            Date = Parsers.ParseDateHeader(response.Headers[ProtocolConstants.Headers.OperationDate]);
             ContentMD5 = response.Headers[ProtocolConstants.Headers.ContentMD5];
         }
